singular dish is now saved instead of being deleted when is present in any order

+++ src/main/java/pl/Aevise/SupperSpeed/business/DishService.java
@@ -4,12 +4,15 @@ import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import pl.Aevise.SupperSpeed.api.dto.DishCategoryDTO;
 import pl.Aevise.SupperSpeed.api.dto.DishDTO;
+import pl.Aevise.SupperSpeed.api.dto.mapper.DishMapper;
 import pl.Aevise.SupperSpeed.business.dao.DishDAO;
 import pl.Aevise.SupperSpeed.domain.Dish;
 import pl.Aevise.SupperSpeed.domain.Image;
 import pl.Aevise.SupperSpeed.infrastructure.database.entity.DishEntity;
 
+import java.util.HashMap;
 import java.util.List;
 import java.util.Optional;
 
@@ -21,6 +24,9 @@ public class DishService {
     private final DishDAO dishDAO;
     private final RestaurantService restaurantService;
     private final DishCategoryService dishCategoryService;
+
+    private final DishListService dishListService;
+    private final DishMapper dishMapper;
 
     public List<Dish> findAllByRestaurant(Integer restaurantId) {
         List<Dish> dishesFromRestaurant = dishDAO.findAllByRestaurant(restaurantId);
@@ -40,8 +46,17 @@ public class DishService {
     }
 
     public void deleteDish(Integer dishId) {
-        dishDAO.deleteDish(dishId);
-        log.info("Deleted dish [{}]", dishId);
+        boolean dishUsed = dishListService.isDishInOrder(dishId);
+
+        if(dishUsed){
+            Optional<Dish> dish = dishDAO.permanentlyHideDishFromAllUsers(dishId);
+            if(dish.isPresent() && dish.get().getIsHidden()){
+                log.info("Dish [{}] hidden from the users", dishId);
+            }
+        }else {
+            dishDAO.deleteDish(dishId);
+            log.info("Deleted dish [{}]", dishId);
+        }
     }
 
     public void deleteDishes(List<DishEntity> dishes) {
@@ -73,5 +88,47 @@ public class DishService {
     public void setDishImage(Image image, Integer dishId) {
         dishDAO.setDishImage(image, dishId);
         log.info("Dish with id: [{}] image updated successfully", dishId);
+    }
+
+//    @Transactional
+//    public HashMap<String, List<DishDTO>> getDishListByCategoryFromRestaurant(Integer restaurantId) {
+//        List<DishCategoryDTO> dishCategories = getDishCategoriesByRestaurantId(restaurantId);
+//        return extractDishesByCategoryName(dishCategories);
+//    }
+
+    @Transactional
+    public HashMap<String, List<DishDTO>> extractDishesByCategoryName(List<DishCategoryDTO> dishCategories) {
+        HashMap<String, List<DishDTO>> dishesByCategory = new HashMap<>();
+
+        for (DishCategoryDTO dishCategory : dishCategories) {
+            dishesByCategory.put(
+                    dishCategory.getCategoryName(),
+                    findAllByCategory(dishCategory.getDishCategoryId())
+                            .stream()
+                            .map(dishMapper::mapToDTO)
+                            .toList()
+            );
+        }
+        return dishesByCategory;
+    }
+
+    @Transactional
+    public HashMap<List<DishCategoryDTO>, List<DishDTO>> extractDishesByCategory(List<DishCategoryDTO> dishCategories, boolean filterUnavailable) {
+        HashMap<List<DishCategoryDTO>, List<DishDTO>> dishesByCategory = new HashMap<>();
+
+        for (DishCategoryDTO dishCategory : dishCategories) {
+            dishesByCategory.put(
+                    List.of(dishCategory),
+                    findAllByCategory(dishCategory.getDishCategoryId())
+                            .stream()
+                            .map(dishMapper::mapToDTO)
+                            .filter(dishDTO -> !filterUnavailable || dishDTO.getAvailability())
+                            .toList()
+            );
+        }
+        if (filterUnavailable) {
+            dishesByCategory.entrySet().removeIf(category -> category.getValue().isEmpty());
+        }
+        return dishesByCategory;
     }
 }
